Add repository interface for session
